Move scaled blas utils back to native and inline functions
